requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-autoslug==1.9.9
django-cities-light==3.10.2
idna==3.11
pillow==12.0.0
progressbar2==4.5.0
python-dotenv==1.2.1
python-utils==3.9.1
pytz==2025.2
razorpay==2.0.0
reportlab==4.4.4
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
